QPACK: fix visibilities
