+++ quinn-h3/src/qpack/mod.rs
@@ -14,6 +14,7 @@ mod field;
 mod static_;
 mod stream;
 mod vas;
+mod parse_error;
 
 mod decoder;
 mod encoder;
@@ -23,27 +24,3 @@ mod prefix_string;
 
 #[cfg(test)]
 mod tests;
-
-#[derive(Debug, PartialEq)]
-pub(self) enum ParseError {
-    InvalidInteger(prefix_int::Error),
-    InvalidString(prefix_string::Error),
-    InvalidPrefix(u8),
-    InvalidBase(isize),
-}
-
-impl From<prefix_int::Error> for ParseError {
-    fn from(e: prefix_int::Error) -> Self {
-        match e {
-            e => ParseError::InvalidInteger(e),
-        }
-    }
-}
-
-impl From<prefix_string::Error> for ParseError {
-    fn from(e: prefix_string::Error) -> Self {
-        match e {
-            e => ParseError::InvalidString(e),
-        }
-    }
-}
