+++ quinn-h3/src/qpack/parse_error.rs
@@ -1,0 +1,25 @@
+use super::{prefix_int, prefix_string};
+
+#[derive(Debug, PartialEq)]
+pub enum ParseError {
+    InvalidInteger(prefix_int::Error),
+    InvalidString(prefix_string::Error),
+    InvalidPrefix(u8),
+    InvalidBase(isize),
+}
+
+impl From<prefix_int::Error> for ParseError {
+    fn from(e: prefix_int::Error) -> Self {
+        match e {
+            e => ParseError::InvalidInteger(e),
+        }
+    }
+}
+
+impl From<prefix_string::Error> for ParseError {
+    fn from(e: prefix_string::Error) -> Self {
+        match e {
+            e => ParseError::InvalidString(e),
+        }
+    }
+}
